requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.15.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
apify_client==1.9.2
apify_shared==1.3.0
appdirs==1.4.4
arxiv==2.1.3
asgiref==3.8.1
asttokens
astunparse==1.6.3
attrs==24.3.0
auth0-python==4.8.1
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.0.2
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
cassandra-driver==3.29.2
cassio==0.1.10
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cloudpathlib==0.20.0
cohere==5.14.0
colorama
coloredlogs==15.0.1
comm 
confection==0.1.5
contourpy==1.3.1
crewai==0.102.0
crewai-tools==0.36.0
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.10
dataclasses-json==0.6.7
debugpy
decorator
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
duckduckgo_search==7.4.4
durationpy==0.9
embedchain==0.1.127
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup 
executing 
faiss-cpu==1.10.0
fastapi==0.115.6
fastavro==1.10.0
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
geojson==2.5.0
geomet==0.2.1.post1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.83.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.2
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
graphviz==0.20.3
greenlet==3.1.1
groq==0.13.1
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata 
importlib_resources==6.4.5
iniconfig==2.0.0
instructor==1.7.3
ipykernel 
ipython 
jaraco.context==6.0.1
jax==0.4.37
jaxlib==0.4.36
jedi 
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
json_repair==0.39.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client 
jupyter_core 
keras==3.7.0
kiwisolver==1.4.8
kubernetes==31.0.0
lancedb==0.20.0
langchain==0.3.13
langchain-chroma==0.1.4
langchain-cohere==0.3.5
langchain-community==0.3.13
langchain-core==0.3.41
langchain-deepseek==0.1.2
langchain-experimental==0.3.4
langchain-google-community==2.0.7
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-mistralai==0.2.7
langchain-nvidia-ai-endpoints==0.3.9
langchain-ollama==0.2.3
langchain-openai==0.2.14
langchain-pinecone==0.2.3
langchain-tests==0.3.13
langchain-text-splitters==0.3.4
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langsmith==0.1.147
language_data==1.3.0
libclang==18.1.1
litellm==1.60.2
lxml==5.3.0
Mako==1.3.9
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.1
matplotlib-inline 
mdurl==0.1.2
mem0ai==0.1.66
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
mxnet==1.2.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.28.0
nest_asyncio 
networkx==3.4.2
nltk==3.9.1
nodeenv==1.9.1
numexpr==2.10.2
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.7
onnxruntime==1.20.1
openai==1.65.4
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
overrides==7.7.0
packaging 
pandas==2.2.3
parso 
pdfminer.six==20231228
pdfplumber==0.11.5
pickleshare 
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs 
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.4
preshed==3.0.9
prompt_toolkit 
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.1
psutil 
psycopg2-binary==2.9.10
pure_eval 
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments 
PyJWT==2.10.1
pylance==0.23.2
pyowm==3.3.0
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.396
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil 
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-oxmsg==0.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==307
PyYAML==6.0.2
pyzmq 
qdrant-client==1.13.3
RapidFuzz==3.11.0
rapidocr-onnxruntime==1.4.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.0.7
shellingham==1.5.4
six 
smart-open==7.0.5
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.0
stack_data 
starlette==0.41.3
streamlit==1.42.2
sympy==1.13.1
syrupy==4.8.2
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
textblob==0.18.0.post0
tf_keras==2.18.0
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.7.0
together==1.4.1
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado 
tqdm==4.67.1
traitlets 
transformers==4.47.0
typer==0.15.1
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions 
tzdata==2025.1
unstructured==0.16.11
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.2.3
uv==0.6.5
uvicorn==0.34.0
validators==0.34.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.3
wcwidth
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wikipedia==1.4.0
wolframalpha==5.1.3
wrapt==1.17.0
xmltodict==0.14.2
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp 
